Refactor compare runner finalizer
